+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-alpha.12] - 2024-11-28
+
+### Added
+- Blog API endpoints:
+  - List blog posts
+  - List blog categories
+  - Get posts by category
+  - Get single post by ID
+  - Get post by slug
+  - Get category by slug with posts
+
 ## [1.0.0-alpha.11] - 2024-11-26
 
 ### Added